requirements.txt
Misskey.py==4.1.0
openai==0.27.8
requests==2.28.1
websockets==11.0.3
